feat: moved the course's teams to a seperate view; UI tweaks
