+++ src/components/DataTable.tsx
@@ -20,13 +20,14 @@ import {
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
 
-type DataTableProps<TData, TValue> = {
+interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[];
   data: TData[];
   onRowClick?: (row: Row<TData>) => void;
   paginationEnabled?: boolean;
   pageSize?: number;
-};
+  showHeader?: boolean; // Add this prop
+}
 
 const DataTable = <TData, TValue>({
   data,
@@ -34,6 +35,7 @@ const DataTable = <TData, TValue>({
   onRowClick,
   pageSize = 10,
   paginationEnabled = false,
+  showHeader = true, // Default to true for backward compatibility
 }: DataTableProps<TData, TValue>) => {
   const { t } = useTranslation();
   const [sorting, setSorting] = React.useState<SortingState>([]);
@@ -62,24 +64,26 @@ const DataTable = <TData, TValue>({
     <div>
       <div className="rounded-md border">
         <Table>
-          <TableHeader>
-            {table.getHeaderGroups().map((headerGroup) => (
-              <TableRow key={headerGroup.id}>
-                {headerGroup.headers.map((header) => {
-                  return (
-                    <TableHead key={header.id}>
-                      {header.isPlaceholder
-                        ? null
-                        : flexRender(
-                            header.column.columnDef.header,
-                            header.getContext()
-                          )}
-                    </TableHead>
-                  );
-                })}
-              </TableRow>
-            ))}
-          </TableHeader>
+          {showHeader && (
+            <TableHeader>
+              {table.getHeaderGroups().map((headerGroup) => (
+                <TableRow key={headerGroup.id}>
+                  {headerGroup.headers.map((header) => {
+                    return (
+                      <TableHead key={header.id}>
+                        {header.isPlaceholder
+                          ? null
+                          : flexRender(
+                              header.column.columnDef.header,
+                              header.getContext()
+                            )}
+                      </TableHead>
+                    );
+                  })}
+                </TableRow>
+              ))}
+            </TableHeader>
+          )}
           <TableBody>
             {table.getRowModel().rows?.length ? (
               table.getRowModel().rows.map((row) => (
