+++ src/components/Modals/CreateTeamModal.tsx
@@ -103,8 +103,9 @@ const CreateTeamModal = ({courseId}: CreateTeamModalProps) => {
   return (
     <Dialog open={open} onOpenChange={setOpen}>
       <DialogTrigger asChild>
-        <Button variant="ghost" size="icon">
+        <Button variant="secondary">
           <Plus className="h-4 w-4"/>
+          {t("coursePageB.teamsTable.createTeam")}
         </Button>
       </DialogTrigger>
       <DialogContent>
